+++ src/domain/flag/handlers.rs
@@ -1,5 +1,5 @@
 use anyhow::Result;
-use himalaya_lib::Backend;
+use himalaya_lib::{Backend, Flags};
 
 use crate::printer::Printer;
 
@@ -7,10 +7,10 @@ pub fn add<P: Printer, B: Backend + ?Sized>(
     printer: &mut P,
     backend: &mut B,
     folder: &str,
-    id: &str,
-    flags: &str,
+    ids: Vec<&str>,
+    flags: &Flags,
 ) -> Result<()> {
-    backend.add_flags(folder, id, flags)?;
+    backend.add_flags(folder, ids, flags)?;
     printer.print("Flag(s) successfully added!")
 }
 
@@ -18,10 +18,10 @@ pub fn set<P: Printer, B: Backend + ?Sized>(
     printer: &mut P,
     backend: &mut B,
     folder: &str,
-    id: &str,
-    flags: &str,
+    ids: Vec<&str>,
+    flags: &Flags,
 ) -> Result<()> {
-    backend.set_flags(folder, id, flags)?;
+    backend.set_flags(folder, ids, flags)?;
     printer.print("Flag(s) successfully set!")
 }
 
@@ -29,9 +29,9 @@ pub fn remove<P: Printer, B: Backend + ?Sized>(
     printer: &mut P,
     backend: &mut B,
     folder: &str,
-    id: &str,
-    flags: &str,
+    ids: Vec<&str>,
+    flags: &Flags,
 ) -> Result<()> {
-    backend.remove_flags(folder, id, flags)?;
+    backend.remove_flags(folder, ids, flags)?;
     printer.print("Flag(s) successfully removed!")
 }